Include conversion of text files to wav
The Saleae Logic 2 program allows us to generate
text (csv) files of the I2S stream from the adapted
MobiWAN. This change includes conversion of this file
into a .wav file and its processing.

Also fixed a bug which made our VAD output look like
it was switching on and off for a few frames at the
end of an utterance.